Write local repo mounts with Compose bind syntax
Generate setup-wizard Docker Compose overrides with long-form bind mounts so local repository paths containing spaces or Compose-sensitive characters are parsed reliably.\n\nConstraint: Local repository setup writes docker-compose.override.yml directly from host paths.\nRejected: Quote short-syntax volume strings | long-form bind mounts avoid source/target delimiter ambiguity more directly.\nConfidence: high\nScope-risk: narrow\nDirective: Keep setup wizard compose output testable through pure helpers.\nTested: node .yarn/releases/yarn-4.7.0.cjs workspace setup-sourcebot test; node .yarn/releases/yarn-4.7.0.cjs workspace setup-sourcebot build\nNot-tested: Full interactive setup-sourcebot wizard flow
