+++ packages/setupWizard/src/dockerCompose.test.ts
@@ -1,0 +1,24 @@
+import assert from 'node:assert/strict';
+import { buildDockerComposeOverride } from './dockerCompose.js';
+
+const override = buildDockerComposeOverride(new Map([
+    ['/Users/sourcebot/Code Projects/repo one', 1],
+    ['/Users/sourcebot/repos/repo:two#main', 0],
+]));
+
+assert.equal(override, [
+    '# Generated by setup-sourcebot',
+    '# Merged with docker-compose.yml at `docker compose up` time.',
+    'services:',
+    '  sourcebot:',
+    '    volumes:',
+    '      - type: bind',
+    '        source: "/Users/sourcebot/repos/repo:two#main"',
+    '        target: /repos/0',
+    '        read_only: true',
+    '      - type: bind',
+    '        source: "/Users/sourcebot/Code Projects/repo one"',
+    '        target: /repos/1',
+    '        read_only: true',
+    '',
+].join('\n'));
