+++ packages/setupWizard/src/dockerCompose.ts
@@ -1,0 +1,24 @@
+const toYamlQuotedString = (value: string): string => {
+    return JSON.stringify(value);
+};
+
+export const buildDockerComposeOverride = (localRepoIndex: Map<string, number>): string => {
+    const mounts = [...localRepoIndex.entries()]
+        .sort((a, b) => a[1] - b[1])
+        .flatMap(([source, index]) => [
+            '      - type: bind',
+            `        source: ${toYamlQuotedString(source)}`,
+            `        target: /repos/${index}`,
+            '        read_only: true',
+        ]);
+
+    return [
+        '# Generated by setup-sourcebot',
+        '# Merged with docker-compose.yml at `docker compose up` time.',
+        'services:',
+        '  sourcebot:',
+        '    volumes:',
+        ...mounts,
+        '',
+    ].join('\n');
+};
