+++ packages/setupWizard/src/index.ts
@@ -16,6 +16,7 @@ import { collectGitHubConfig } from './github.js';
 import { collectGitLabConfig } from './gitlab.js';
 import { collectLocalReposConfig } from './localRepos.js';
 import { collectModels, PROVIDER_ENV_KEYS } from './models.js';
+import { buildDockerComposeOverride } from './dockerCompose.js';
 import {
     type CollectResult,
     type ConnectionConfig,
@@ -616,19 +617,7 @@ async function main() {
     const writtenFiles = ['config.json', '.env'];
 
     if (localRepoIndex.size > 0) {
-        const mounts = [...localRepoIndex.entries()]
-            .sort((a, b) => a[1] - b[1])
-            .map(([p, i]) => `      - ${p}:/repos/${i}:ro`);
-        const overrideYaml = [
-            '# Generated by setup-sourcebot',
-            '# Merged with docker-compose.yml at `docker compose up` time.',
-            'services:',
-            '  sourcebot:',
-            '    volumes:',
-            ...mounts,
-            '',
-        ].join('\n');
-        writeFileSync('docker-compose.override.yml', overrideYaml);
+        writeFileSync('docker-compose.override.yml', buildDockerComposeOverride(localRepoIndex));
         writtenFiles.push('docker-compose.override.yml');
     }
